pyproject.toml
[project]
name = "blob-dict"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]

[project.optional-dependencies]
image = [
    "pillow>=11.1.0",
]

[dependency-groups]
dev = [
    "ipython>=9.0.2",
    "mypy>=1.15.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = "--cov --cov-report xml:cov.xml --cov-report html --cov-report term-missing test"
pythonpath = [
    "src",
]

[tool.coverage.run]
source = [
    "src",
]
